requirements.txt
numpy>=1.21.0
Pillow>=9.0.0
pynput>=1.7.6
pytesseract>=0.3.10
opencv-python>=4.8.0
